+++ src/main/java/com/construccion/repository/ParametrosRepository.java
@@ -3,5 +3,5 @@ package com.construccion.repository;
 import com.construccion.entity.ParametrosConstruccion;
 
 public interface ParametrosRepository {
-	ParametrosConstruccion consultarParametrosPorConstruccion(Long idConstruccion);
+	ParametrosConstruccion consultarParametrosPorConstruccion(Integer idConstruccion);
 }
